+++ src/references/index.ts
@@ -1,6 +1,7 @@
 import { mapKeys, mapValues, toLower } from 'lodash';
 import savingAssets from './compound/saving-assets.json';
 import tokenOverridesData from './token-overrides.json';
+import { Asset } from '@rainbow-me/entities';
 
 export { default as chains } from './chains.json';
 export { default as compoundCERC20ABI } from './compound/compound-cerc20-abi.json';
@@ -9,6 +10,13 @@ export { default as emojis } from './emojis.json';
 export { default as erc20ABI } from './erc20-abi.json';
 export { default as ethUnits } from './ethereum-units.json';
 export { default as shitcoins } from './shitcoins.json';
+
+interface OverrideInfo {
+  color?: string;
+  name?: string;
+  shadowColor?: string;
+  symbol?: string;
+}
 
 export const CDAI_CONTRACT = '0x5d3a536e4d6dbd6114cc1ead35777bab948e3643';
 export const SAI_ADDRESS = '0x89d24a6b4ccb1b6faa2625fe562bdd9a23260359';
@@ -68,15 +76,16 @@ export const DefaultUniswapFavorites = {
   ],
 };
 
-export const tokenOverrides = mapKeys(tokenOverridesData, (_, address) =>
-  toLower(address)
-);
+export const tokenOverrides: Record<
+  string,
+  OverrideInfo
+> = mapKeys(tokenOverridesData, (_, address) => toLower(address));
 
 export const savingsAssetsList = savingAssets;
 
 export const savingsAssetsListByUnderlying = mapValues(
   savingAssets,
-  assetsByNetwork =>
+  (assetsByNetwork: Record<string, Asset>) =>
     mapKeys(
       mapValues(assetsByNetwork, (assetByContract, contractAddress) => ({
         ...assetByContract,