+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Removed
 
+## [1.2.39](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.39)
+### Added
+* Surface Uniswap V2 LP tokens
+* Add swap search spinner
+
 ## [1.2.38](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.38)
 ### Added
 * Uniswap V2 support